Simplify how inter-package script dependencies are named.
E.g. if a package named "foo" is installed, then any script can simply
do "@load foo" or "@load foo/some_script".  So specifying script
dependencies between packages is easy/consistent and such packages don't
necessarily require installation via package-manager to work.

+++ bropkg/util.py
@@ -61,7 +61,7 @@ def make_symlink(target_path, link_path, force=True):
     try:
         os.symlink(target_path, link_path)
     except OSError as error:
-        if error.errno == errno.EEXIST and force:
+        if error.errno == errno.EEXIST and force and os.path.islink(link_path):
             os.remove(link_path)
             os.symlink(target_path, link_path)
         else:
